+++ includes/object-cache.php
@@ -1379,7 +1379,14 @@ LUA;
         $start_time = microtime( true );
 
         try {
-            $results = array_combine( $remaining_keys, $this->redis->mget( $remaining_keys ) );
+            $results = array_combine(
+                $remaining_keys,
+                $this->redis->mget(
+                    array_map( function ( $key ) use ( $derived_keys ) {
+                        return $derived_keys[ $key ];
+                    }, $remaining_keys )
+                )
+            );
         } catch ( Exception $exception ) {
             $this->handle_exception( $exception );